tests/kernel: queue characters through the real tty_subr.c
Every tty queues its input and output through the character lists in
sys/kern/tty_subr.c, and nothing ran them off the board. The gate links the
kernel source and pins the boundary cases the block chain turns on: a block
that fills exactly, one that empties exactly from either end, a free list
that runs out mid-transfer, the contiguous run ndqb reports, the walk nextc
makes without consuming, and catq taking a chain outright when its
destination is empty. 321 assertions.

Two things had to change first.

tty_subr.c cast character pointers through int in twenty places, which is
correct only where a pointer fits one. Those are now u_long, which is
pointer-sized on the port and on an LP64 host alike, and two sites became
plain pointer arithmetic. Cross-compiled with the kernel's own flags the
object is the same 1352 bytes of text over the same 639 instructions;
register allocation shifts inside ndqb, where two statements merged into
one. init_main.c's cinit() carried the same cast and is byte-identical
after the same treatment. Neither claim covers stm32 or pic32, which
consume both files and which this tree has no toolchain for.

ndqb and ndflush had no prototype anywhere and no caller outside
tty_subr.c. They now sit in <sys/tty.h> beside the other seven clist
entry points, inside the KERNEL guard the rest of that block is in, so
usr.bin/w and usr.sbin/pstat see nothing new.

The harness gains an interrupt layer: the port reaches PRIMASK through ARM
inline assembly, so the generated <machine/intr.h> takes the port's header
for its constants and then hostintr.h, which redefines the six spl macros.
It counts the level rather than discarding it, so every scenario also
states that the routine lowered priority again -- one that did not would
wedge the console rather than fail a comparison.

macOS CI caught two clang diagnostics gcc does not raise: old-style
definitions in sys/kern, and operand widths in asm that is never emitted.
Both are excused for this tier the way exec_spool, cache_footprint and
bin/sh/tests already excuse them.

+++ sys/kern/init_main.c
@@ -259,11 +259,11 @@ binit(void)
 static void
 cinit(void)
 {
-	int ccp;
+	u_long ccp;
 	struct cblock *cp;
 
-	ccp = (int)cfree;
-	ccp = (ccp + CROUND) & ~CROUND;
+	ccp = (u_long)cfree;
+	ccp = (ccp + CROUND) & ~(u_long)CROUND;
 	for (cp = (struct cblock *)ccp; cp <= &cfree[NCLIST - 1]; cp++) {
 		cp->c_next = cfreelist;
 		cfreelist = cp;
